Commit Mojolicious::Lite proof of concept
This is a Mojolicious::Lite application which can be run under the
morbo development server. There is a Local::MockLibraryREST module
which can be used for testing and design work before beginning on
Evergreen and/or Koha implementation modules.

This pluggable implementation method is inspired by SIPServer.

In theory, this application could be run under mod_perl on an
Evergreen system, or run under Plack on a Koha system.

Many thanks to everyone at the Evergreen 2013 conference who
participated in discussion and design of this idea, but at the
hackfest and in various hallway discussions.

Signed-off-by: Jeff Godin <jgodin@tadl.org>